Add factory functions and storage functions
So im beginning the project. Already done two
factory functions, one for projects and another
for tasks, on the to-do.js file.

In the storage.js file, i code three functions:
one for storage data, other for delete that data, and
anotherone to retrieve data. This three can be
organized in an object.