Change project element function
Same as with the task function, i splitted
the project function into three:
The first is the project element creator, which
creates the element.
The second is the project element added, which
add the created element in the dom.
the third is home render, which is going to
retrieve and render the projects from localStorage.